On-the-fly static rendering of IPython notebooks
If enabled (via the ipython_notebook.render configuration setting),
visiting the blob view on any .ipynb file triggers a call to IPython
nbconvert and the result is included in an iframe. IPython (>= 1.0)
must be installed on the server.

Note that this is a very ad-hoc patch. For example, rendered notebooks
are not cached meaning that performance may be suboptimal.